python: Upgrade pip just before using it
